requirements.txt
selenium
pandas
PyQt6
openpyxl
mysql-connector-python